+++ compiler/ml/tast_mapper.ml
@@ -182,7 +182,8 @@ let pat sub x =
 let expr sub x =
   let extra = function
     | Texp_constraint cty -> Texp_constraint (sub.typ sub cty)
-    | Texp_coerce cty2 -> Texp_coerce (sub.typ sub cty2)
+    | Texp_coerce {source_type; target; target_type} ->
+      Texp_coerce {source_type; target = sub.typ sub target; target_type}
     | Texp_open (ovf, path, loc, env) ->
       Texp_open (ovf, path, loc, sub.env sub env)
   in